Refractor ai route to separate files
